# openapi.yml
...
openapi: 3.0.3
info:
  title: sombreroquark API
  version: 1.0.0-SNAPSHOT
...
paths:
...
  /api/users:
    post:
      tags:
      - Users
      parameters:
      - name: redirect
        in: query
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDTO'
      responses:
        "200":
          description: User created
          content:
            application/json:
              schema:
                type: object
              example:
                id: 7883e4aa-1978-46f0-9712-c468cc7dd87c
                username: jhon
                email: doe@smail.com
                createdAt: 2023-03-26T22:55:36.900710700Z
                active: true
                emailVerified: false
                data:
                  note: here you can enter extra data for user registration
                  name: jhon
                groups: []
        "400":
          description: User already exists
          content:
            application/json:
              schema:
                type: object
              example:
                error: User already exists
                code: 400
  /api/users/findUserByEmail:
    post:
      tags:
      - Users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialsDTO'
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                type: object
              example:
                id: 7883e4aa-1978-46f0-9712-c468cc7dd87c
                username: jhon
                email: doe@smail.com
                createdAt: 2023-03-26T22:55:36.900710700Z
                active: true
                emailVerified: false
                data:
                  note: here you can enter extra data for user registration
                  name: jhon
                groups: []
        "401":
          description: Unauthorized
        "400":
          description: Missing permissions
          content:
            application/json:
              schema:
                type: object
                example:
                  error: Missing permissions
                  code: 400
        "404":
          description: User not found
          content:
            application/json:
              schema:
                type: object
                example:
                  error: User not found
                  code: 404
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /api/users/getPaginatedUsers:
    post:
      tags:
      - Users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaginatedRequestDTO'
      responses:
        "200":
          description: Paginated users
          content:
            application/json:
              schema:
                type: object
              example:
                users:
                - id: 7883e4aa-1978-46f0-9712-c468cc7dd87c
                  username: jhon
                  email: doe@smail.com
                  createdAt: 2023-03-26T22:55:36.900710Z
                  active: true
                  emailVerified: false
                  data:
                    name: jhon
                    note: here you can enter extra data for user registration
                  groups: []
                currentPage: 1
                pageSize: 10
                totalItems: 3
        "401":
          description: Unauthorized
        "400":
          description: Missing permissions
          content:
            application/json:
              schema:
                type: object
                example:
                  error: Missing permissions
                  code: 400
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /api/users/me:
    post:
      tags:
      - Users
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                type: object
              example:
                id: 7883e4aa-1978-46f0-9712-c468cc7dd87c
                username: jhon
                email: doe@smail.com
                createdAt: 2023-03-26T22:55:36.900710700Z
                active: true
                emailVerified: false
                data:
                  note: here you can enter extra data for user registration
                  name: jhon
                groups: []
        "401":
          description: Unauthorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /api/users/me/changepassword:
    post:
      tags:
      - Users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialsDTO'
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                type: object
              example:
                id: 7883e4aa-1978-46f0-9712-c468cc7dd87c
                username: jhon
                email: doe@smail.com
                createdAt: 2023-03-26T22:55:36.900710700Z
                active: true
                emailVerified: false
                data:
                  note: here you can enter extra data for user registration
                  name: jhon
                groups: []
        "401":
          description: Unauthorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /api/users/me/updatedata:
    post:
      tags:
      - Users
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: object
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                type: object
              example:
                id: 7883e4aa-1978-46f0-9712-c468cc7dd87c
                username: jhon
                email: doe@smail.com
                createdAt: 2023-03-26T22:55:36.900710700Z
                active: true
                emailVerified: false
                data:
                  note: here you can enter extra data for user registration
                  name: jhon
                groups: []
        "401":
          description: Unauthorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /api/users/reset/password:
    get:
      tags:
      - Users
      parameters:
      - name: redirect
        in: query
        schema:
          type: string
      - name: token
        in: query
        schema:
          type: string
      responses:
        "200":
          description: Email sent
          content:
            application/json:
              schema:
                type: object
              example:
                message: Email sent
                code: 200
        "400":
          description: Token already used
          content:
            application/json:
              schema:
                type: object
                example:
                  error: Token already used
                  code: 400
    put:
      tags:
      - Users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialsDTO'
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                type: object
              example:
                message: Reset password successfully
                code: 200
        "400":
          description: User not found
          content:
            application/json:
              schema:
                type: object
                example:
                  error: User not found
                  code: 400
    post:
      tags:
      - Users
      parameters:
      - name: email
        in: query
        schema:
          type: string
      - name: redirect
        in: query
        schema:
          type: string
      responses:
        "200":
          description: Email sent
          content:
            application/json:
              schema:
                type: object
              example:
                message: Email sent
                code: 200
        "400":
          description: Redirect url not allowed
          content:
            application/json:
              schema:
                type: object
                example:
                  error: Redirect url not allowed
                  code: 400
  /api/users/updatedata/{email}:
    post:
      tags:
      - Users
      parameters:
      - name: email
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: object
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                type: object
              example:
                id: 7883e4aa-1978-46f0-9712-c468cc7dd87c
                username: jhon
                email: doe@smail.com
                createdAt: 2023-03-26T22:55:36.900710700Z
                active: true
                emailVerified: false
                data:
                  note: here you can enter extra data for user registration
                  name: jhon
                groups: []
        "401":
          description: Unauthorized
        "400":
          description: User not found
          content:
            application/json:
              schema:
                type: object
                example:
                  error: User not found
                  code: 400
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /api/users/verify/email:
    get:
      tags:
      - Users
      parameters:
      - name: redirect
        in: query
        schema:
          type: string
      - name: token
        in: query
        schema:
          type: string
      responses:
        "302":
          description: Redirect to url
        "400":
          description: Redirect url not valid
          content:
            application/json:
              schema:
                type: object
              example:
                error: Redirect url not valid
                code: 400
...
components:
  schemas:
...
    CreateUserDTO:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        data:
          type: object
          additionalProperties:
            type: object
        emailVerified:
          type: boolean
    CredentialsDTO:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        newPassword:
          type: string
...
    PaginatedRequestDTO:
      type: object
      properties:
        page:
          format: int32
          type: integer
        pageSize:
          format: int32
          type: integer
...
  securitySchemes:
    SecurityScheme:
      type: http
      description: Authentication
      scheme: bearer
      bearerFormat: JWT